Move migrations to a separate directory
