+++ storage/migration/cadence_values_migration.go
@@ -1,11 +1,28 @@
-package sqlite
+/*
+ * Flow Emulator
+ *
+ * Copyright 2024 Dapper Labs, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package migration
 
 import (
 	"context"
 	"database/sql"
 	"encoding/binary"
 	"encoding/hex"
-	"fmt"
 	"os"
 	"strconv"
 	"strings"
@@ -13,6 +30,7 @@ import (
 	"github.com/rs/zerolog"
 
 	"github.com/onflow/flow-emulator/storage"
+	"github.com/onflow/flow-emulator/storage/sqlite"
 
 	"github.com/onflow/cadence/runtime/common"
 	"github.com/onflow/cadence/runtime/interpreter"
@@ -24,13 +42,8 @@ import (
 	"github.com/onflow/flow-go/model/flow"
 )
 
-func Migrate(url string) error {
-	store, err := New(url)
-	if err != nil {
-		return err
-	}
-
-	payloads, payloadInfo, accounts, err := payloadsAndAccountsFromSnapshot(store.db)
+func MigrateCadenceValues(store *sqlite.Store) error {
+	payloads, payloadInfo, accounts, err := payloadsAndAccountsFromSnapshot(store.DB())
 	if err != nil {
 		return err
 	}
@@ -125,7 +138,7 @@ func newConsoleLogger() zerolog.Logger {
 }
 
 func writePayloadsToSnapshot(
-	store *Store,
+	store *sqlite.Store,
 	payloads []*ledger.Payload,
 	payloadInfoSet map[flow.RegisterID]payloadMetaInfo,
 ) error {
@@ -149,18 +162,17 @@ func writePayloadsToSnapshot(
 
 		payloadInfo, ok := payloadInfoSet[registerId]
 		if ok {
-			_, err = store.db.Exec(
-				fmt.Sprintf(
-					"INSERT INTO %s (key, version, value, height) VALUES (?, ?, ?, ?) ON CONFLICT(key, version, height) DO UPDATE SET value=excluded.value",
-					storeName,
-				),
-				hex.EncodeToString(registerIdBytes),
+			// Insert the values back with the existing height and version.
+			err = store.SetBytesWithVersionAndHeight(
+				nil,
+				storeName,
+				registerIdBytes,
+				value,
 				payloadInfo.version,
-				hex.EncodeToString(value),
 				payloadInfo.height,
 			)
 		} else {
-			// if this is a new payload, use the current block height
+			// If this is a new payload, use the current block height, and default version.
 			err = store.SetBytes(
 				nil,
 				storeName,
@@ -196,7 +208,7 @@ func payloadsAndAccountsFromSnapshot(db *sql.DB) (
 
 	for rows.Next() {
 		var hexKey, hexValue string
-		var height, version int
+		var height, version uint64
 
 		err := rows.Scan(&hexKey, &hexValue, &height, &version)
 		if err != nil {
@@ -282,7 +294,7 @@ func registerIDKeyFromString(s string) (flow.RegisterID, common.Address) {
 }
 
 type payloadMetaInfo struct {
-	height, version int
+	height, version uint64
 }
 
 type ReportWriterFactory struct{}
