+++ storage/migration/test-data/README.md
@@ -1,0 +1,3 @@
+This is the same emulator state used in https://github.com/onflow/flow-go/tree/feature/stable-cadence/cmd/util/ledger/migrations/test-data/cadence_values_migration
+
+Follow the instruction there to generate a new state, if needed.
